Улучшены магазины
В магазинах добавлено отображение:
- Общего количества домов и занятых домов
- Количество защиты.

+++ src/ru/majestic/thetown/game/cargo/ACargo.java
@@ -13,7 +13,7 @@ public abstract class ACargo implements ICargo {
    
    @Override
    public void load(SharedPreferences prefs) {
-      count = prefs.getInt(getSaveTag() + SAVE_SUFFIX_COUNT, 0);
+      count = prefs.getLong(getSaveTag() + SAVE_SUFFIX_COUNT, 0);
    }
 
    @Override
